requirements.txt
# Core Components
streamlit==1.52.0

streamlit-clickable-images==0.0.3

# Computer Vision & AI
insightface==0.7.3
opencv-python-headless==4.10.0.84
scikit-learn==1.5.2

# Math & Data
# Note: Numpy < 2.0 to avoid breaking changes in other libs
numpy==1.26.4
ml-dtypes==0.5.3

# Image Processing
Pillow==12.0.0
PyTurboJPEG>=1.7.0  # Fast JPEG decoding (2-3x faster than PIL)
exifread>=3.0.0     # EXIF orientation handling for TurboJPEG

## ONNX Runtime - if you are not using docker to run program use one of below
#onnxruntime==1.23.2      # CPU image ke liye
onnxruntime-gpu==1.20.1
hdbscan>=0.8.33

# System Monitoring
psutil
pynvml
